requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
get==2019.4.13
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
post==2019.4.13
public==2019.4.13
pylint==2.4.4
python-decouple==3.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
